requirements.txt
Fabric==1.6.0
Flask==0.9
Flask-DebugToolbar==0.8.0
Flask-PyMongo==0.2.1
Jinja2==2.6
PyJWT==0.1.5
Werkzeug==0.8.3
amqp==1.0.11
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.28
blinker==1.2
boto==2.6.0
celery==3.0.19
celery-with-mongodb==3.0
-e git+https://github.com/vvasabi/foursquare.git@daa7e4a1a28fe4a8a0fe687a17b6b33da57ea7f0#egg=foursquare-dev
gunicorn==0.17.4
heroku==0.1.2
httplib2==0.8
itsdangerous==0.17
kombu==2.5.10
newrelic==1.11.0.55
paramiko==1.10.1
poster==0.8.1
pycrypto==2.6
pymongo==2.5
python-dateutil==2.1
python-postmark==0.3.2
raven==3.3.4
requests==1.2.0
simplejson==3.1.3
six==1.3.0
ssh==1.8.0
twilio==3.4.5
unittest2==0.5.1
wsgiref==0.1.2
yuicompressor==2.4.7
